// src/entities/ticketList/types/ITicketProps.ts
export interface ITicketProps {
    data: TicketTypes
}

export type TicketTypes = {
    origin: string;
    origin_name: string;
    destination: string;
    destination_name: string;
    departure_date: string;
    departure_time: string;
    arrival_date: string;
    arrival_time: string;
    carrier: string;
    stops: number;
    price: number;
}